refactor: 게임 진행을 더 할 수 없는 상태들을 추상 클래스 Finished의 상속 관계로 묶어줌
