+++ src/main/java/blackjack/domain/state/Blackjack.java
@@ -1,25 +1,10 @@
 package blackjack.domain.state;
 
-import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
-import java.util.Objects;
 
-public final class Blackjack implements State {
-
-    private final Cards cards;
+public final class Blackjack extends Finished {
 
     Blackjack(Cards cards) {
-        Objects.requireNonNull(cards, "[ERROR] 카드패는 null일 수 없습니다.");
-        this.cards = cards;
-    }
-
-    @Override
-    public State hit(Card card) {
-        throw new UnsupportedOperationException("[ERROR] 게임을 진행할 수 없습니다.");
-    }
-
-    @Override
-    public State stand() {
-        throw new UnsupportedOperationException("[ERROR] 게임을 진행할 수 없습니다.");
+        super(cards);
     }
 }
